requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.3.2
pyttsx3==2.90
Werkzeug==2.3.7
mysql-connector-python
gTTS
assemblyai
langdetect
